redirect to /profile instead of /onboarding, and redirect from /profile to /onboarding if no course history
This is the "right way" to do it because it makes no sense to have someone hop
from /onboarding to /profile on every login for every user. It also lets anyone
go to /onboarding whenever they want naturally. However, this means we can't
skip onboarding. To do that, we should set a flag in user for last skipped date
or something, and if that's less than some hours then we don't redirect from
profile.