Foreground the local server for make local
Stop running the flask server in the background so we can use the python
debugger. With it running in the background, the debugger automatically exits
because it isn't interactive. With this running in the foreground, using the
debugging by injecting

    import ipdb; ipdb.set_trace()

Works fine (you'll need to pip install ipdb).

Ctrl-C will still kill the flask server and all the other local server
processes.