Use SES for sending emails to hotmail users
The change that was necessary to fix this was changing the DNS records:
- spf2.0/pra include:amazonses.com ?all
- v=spf1 include:amazonses.com ?all

to:

- spf2.0/pra include:amazonses.com -all
- v=spf1 include:amazonses.com -all

i.e. "?all" => "-all"

From http://docs.aws.amazon.com/ses/latest/DeveloperGuide/spf.html:

> If you use "-all" as shown in the example above, ISPs may block email
> from IP addresses that are not listed in your SPF record. You
> therefore must add a record for every IP address that you send email
> from. As a debugging aid, you can use "~all" instead. **When you use
> "~all", ISPs will typically accept email from IP addresses that are
> not listed. However, they may flag it**. To maximize deliverability, use
> "-all" and add a record for each IP address.
that are not listed. However, they may flag it.

I'm guessing the same might be happening when using "?all".

http://www.openspf.org/Frank_Ellermann/OAQ explains the difference
between "-all", "?all", and "~all".