Auto scroll to the next course after reviewing
A course is said to be "mostly filled in" when it has at least one ratings for
both the course and professor, as well as both the course and professor reviews.

Auto scroll occurs when either:
  1) A review is saved and the course is "mostly filled in".
  2) A rating is given and the course is "mostly filled in" and all professor
     rating criteria are completely filled in.

Case 2) is a heuristic that tries to pretend when the user would want the auto
scroll feature.

Auto scroll does not occur when editing a course that is already "mostly filled
in".