have deploy script clone to a new dir first, then symlink over
This is so we can have recovery points by preserving the last few git clones.
This also ensures we don't have inconsistent folder state during compiles.

However, this can't be used till we remove dependencies on files in the current
directory. All data files should live in ~/data, all logs in ~/log. Do not write
or store anything in the ~/rmc folder!