Fix schedule share bar to not wrap on small widths
During our CUSEC demo, I noticed that the "export" button overflowed to the next
line. In fact this starts happening on any screen widths < 1300px.

This fixes that, and also gets the link input to stretch to take up all
available space. Also refactors the CSS/HTML to get gutter widths to be more
consistent and have only one (well, two) occurrence of that width.