Add a secrets.py.example and copy that in setup.sh
This is for new developers and open-source contributors so they won't have to
ask us and we don't have to give out our secrets.py.

Test plan: ran `make` and ensured local server still works. Unfortunately, login
with Facebook does not work with a dummy `FB_APP_SECRET_DEV`. New devs will
either have to ask us for that, or we should figure out if it's safe to just
check that in, or figure out some other workaround in dev.