Add /api/v1/login/facebook endpoint
Authenticates clients that provide a Facebook user ID and access token. This is
useful for authenticating mobile apps where Facebook's SDK does not provide
a "signed request" that encodes the FBID and access token, like the web SDK
does.

We respond with a session cookie in a response header that can be used for API
calls that accept user auth.