Remove Sift Science tracking
This was previously added because I interned at siftscience.com and I integrated
Flow as a Sift Science "customer," partly for learning, and partly as a chance
to write our first Python API client. This Flow test customer also proved to be
helpful for internal testing.

I haven't been involved with Sift for a while now, so removing this tracking.