"Search courses" --> "Explore courses"
The main motivation for this change is to make it more likely for someone to
want to click on this nav heading if they just landed on the homepage,
especially if they don't have a particular course in mind to want to search for.

This is exactly the type of thing that would be great to A/B test, but we don't
get enough traffic to yield a meaningful result within a short time frame.